+++ src/Certes/Acme/AcmeDirectory.cs
@@ -99,6 +99,15 @@ namespace Certes.Acme
             /// </value>
             [JsonProperty("caa-identities")]
             public IList<string> CaaIdentities { get; set; }
+
+            /// <summary>
+            /// Indicate if ACME accounts requires external account binding (optional, https://tools.ietf.org/html/rfc8555#section-7.1.1)
+            /// </summary>
+            /// <value>
+            /// If true, requires external account binding
+            /// </value>
+            [JsonProperty("externalAccountRequired")]
+            public bool? ExternalAccountRequired { get; set; }
         }
     }
 }