+++ docs/CHANGELOG.md
@@ -3,8 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
-- Support alternate link relations 
+- Support alternate link relations ([#232][i232])
 - [CLI] Add option for alternate link relation
+- Implement external account binding support (library). ([#231][i231])
 
 ## [2.3.4] - 2020-03-27
 ### Changed